requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.16
django-autoslug==1.9.3
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-extensions==2.1.3
django-filer==1.3.2
django-filter==2.0.0
django-geoposition==0.3.1
django-geosimple==1.0.3
django-js-asset==1.1.0
django-localflavor==2.1
django-mptt==0.9.1
django-polymorphic==2.0.3
django-rest-swagger==2.2.0
django-tastypie==0.14.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
djangorestframework-recursive==0.1.2
easy-thumbnails==2.5
fake-factory==9999.9.9
geopy==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.0.0
psycopg2==2.7.3.2
pyfcm==1.4.5
python-dateutil==2.7.0
python-geohash==0.8.5
python-mimeparse==1.6.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-toolbelt==0.8.0
simplejson==3.13.2
six==1.10.0
sqlparse==0.2.4
stripe==2.10.1
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
